+++ wikidata/entity.py
@@ -3,8 +3,12 @@
 
 """
 import collections.abc
+import datetime
 import enum
-from typing import TYPE_CHECKING, Hashable, Mapping, NewType, Optional, Union
+import logging
+from typing import (TYPE_CHECKING,
+                    Hashable, ItemsView, Iterator, Mapping, NewType, Optional,
+                    Sequence, Tuple, Union, ValuesView)
 
 from babel.core import Locale, UnknownLocaleError
 
@@ -96,9 +100,26 @@ class EntityType(enum.Enum):
     property = 'property'
 
 
-class Entity(Hashable if TYPE_CHECKING else collections.abc.Hashable):
+class Entity(Mapping['Entity', object]
+             if TYPE_CHECKING
+             else collections.abc.Mapping,
+             Hashable if TYPE_CHECKING else collections.abc.Hashable):
     """Wikidata entity.  Can be an item or a property.  Its attrributes
     can be lazily loaded.
+
+    To get an entity use :meth:`Client.get() <wikidata.client.Client.get>`
+    method instead of the constructor of :class:`Entity`.
+
+    .. note::
+
+       Although it implements :class:`~typing.Mapping`\ [:class:`EntityId`,
+       :class:`object`], it actually is multidict.  See also :meth:`getlist()`
+       method.
+
+    .. versionchanged:: 0.2.0
+
+       Implemented :class:`~typing.Mapping`\ [:class:`EntityId`,
+       :class:`object`] protocol for easy access of statement values.
 
     .. versionchanged:: 0.2.0
 
@@ -119,12 +140,119 @@ class Entity(Hashable if TYPE_CHECKING else collections.abc.Hashable):
         if not isinstance(other, type(self)):
             raise TypeError(
                 'expected an instance of {0.__module__}.{0.__qualname__}, '
-                'not {!r}'.format(type(self), other)
+                'not {1!r}'.format(type(self), other)
             )
         return other.id == self.id and self.client is other.client
 
     def __hash__(self) -> int:
         return hash((self.id, id(self.client)))
+
+    def __len__(self) -> int:
+        return len(self.attributes.get('claims', {}))
+
+    def __iter__(self) -> Iterator['Entity']:
+        client = self.client
+        for prop_id in self.attributes.get('claims', {}):
+            yield client.get(prop_id)
+
+    def __getitem__(self, key: 'Entity') -> object:
+        result = self.getlist(key)
+        if result:
+            return result[0]
+        raise KeyError(key)
+
+    def getlist(self, key: 'Entity') -> Sequence[object]:
+        """Return all values associated to the given ``key`` property
+        in sequence.
+
+        :param key: The property entity.
+        :type key: :class:`Entity`
+        :return: A sequence of all values associated to the given ``key``
+                 property.  It can be empty if nothing is associated to
+                 the property.
+        :rtype: :class:`~typing.Sequence`\ [:class:`object`]
+
+        """
+        if not (isinstance(key, type(self)) and
+                key.type is EntityType.property):
+            return []
+        claims = self.attributes.get('claims', {}).get(key.id, [])
+        claims.sort(key=lambda claim: claim['rank'],  # FIXME
+                    reverse=True)
+        logger = logging.getLogger(__name__ + '.Entity.getitem')
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug('claim data: %s',
+                         __import__('pprint').pformat(claims))
+        client = self.client
+
+        def decode(datavalue: Mapping[str, object]):  # FIXME
+            type_ = datavalue['type']
+            value = datavalue['value']
+            if type_ == 'string':
+                return value
+            elif type_ == 'time':
+                cal = value['calendarmodel']
+                if cal == 'http://www.wikidata.org/entity/Q1985727':
+                    time = value['time']
+                    if time[0] != '+':
+                        raise ValueError(
+                            '{!r}: only AD (CE) is supported: {!r}'.format(
+                                time, datavalue
+                            )
+                        )
+                    elif value['timezone'] != 0:
+                        raise ValueError(
+                            '{!r}: timezone other than 0 is unsupported: '
+                            '{!r}'.format(value['timezone'], datavalue)
+                        )
+                    elif value['before'] != 0 or value['after'] != 0:
+                        raise ValueError(
+                            'Uncertainty range time (represented using before/'
+                            'after) is unsupported: ' + repr(datavalue)
+                        )
+                    precision = value['precision']
+                    if precision == 11:
+                        return datetime.date(int(time[1:5]), int(time[6:8]),
+                                             int(time[9:11]))
+                    elif 12 <= precision <= 14:
+                        return datetime.datetime.strptime(
+                            time[1:],
+                            '%Y-%m-%dT%H:%M:%S%z'
+                        )
+                    else:
+                        raise ValueError(
+                            '{!r}: time precision other than 11 to 14 is '
+                            'unsupported: {!r}'.format(precision, datavalue)
+                        )
+                raise ValueError('{!r} is unsupported calendarmodel for time '
+                                 'datavalue: {!r}'.format(cal, datavalue))
+            elif type_ == 'wikibase-entityid':
+                return client.get(value['id'])
+            raise ValueError('{0[type]} is unsupported datavalue '
+                             'type: {0!r}'.format(datavalue))
+        return [decode(claim['mainsnak']['datavalue']) for claim in claims]
+
+    def iterlists(self) -> Iterator[Tuple['Entity', Sequence[object]]]:
+        for prop in self:
+            yield prop, self.getlist(prop)
+
+    def lists(self) -> ItemsView['Entity', Sequence[object]]:
+        """Similar to :meth:`items()` except the returning pairs have
+        each list of values instead of each single value.
+
+        :return: The pairs of (key, values) where values is a sequence.
+        :rtype: :class:`~typing.ItemsView`\ [:class:`Entity`,
+                    :class:`~typing.Sequence`\ [:class:`object`]]
+
+        """
+        return list(self.iterlists())
+
+    def iterlistvalues(self) -> Iterator[Sequence[object]]:
+        for _, values in self.iterlists():
+            yield values
+
+    def listvalues(self) -> ValuesView[Sequence[object]]:
+        return list(self.iterlistvalues())
 
     @property
     def type(self) -> EntityType: